requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
environ==1.0
idna==2.10
multidict==5.1.0
pycodestyle==2.7.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
slackclient==2.9.3
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
